pm2 and artillery, moved json files
