+++ src/test/_init.spec.ts
@@ -3,7 +3,7 @@ process.env.NODE_ENV = 'TEST';
 import * as chai from 'chai';
 import MongooseModels from '../models/mongoose-models';
 import mongoose from '../server/init';
-import * as stixArray from './stix_array.json';
+import * as stixArray from '../../testassets/stix_array.json';
 import config from '../services/config.service';
 
 console.log('Initializing test data');
