+++ testassets/stix.json
