+++ testassets/stix_array.json
